requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-dotenv
pydantic
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt
requests
